Arreglo errores de compilacion con flags -Wall -Werror -Wconversion

+++ administrativo.cpp
@@ -4,13 +4,10 @@
 
 #include "administrativo.h"
 
-Administrativo::Administrativo(uint64_t legajo, std::string nombre, std::string apellido, uint64_t salario, uint64_t antiguedad)
-        :Empleado(legajo, nombre, apellido, salario, antiguedad)
-{
-
-}
+Administrativo::Administrativo(unsigned int legajo, std::string nombre, std::string apellido, float salario, unsigned int antiguedad)
+        :Empleado(legajo, nombre, apellido, salario, antiguedad) {}
 
 void Administrativo::incrementarSalario()
 {
-    this->salario = this->salario * 1.2;
+    this->salario = this->salario * 1.2f;
 }
