+++ jefeDeArea.cpp
@@ -4,7 +4,7 @@
 
 #include "jefeDeArea.h"
 
-JefeDeArea::JefeDeArea(uint64_t legajo, std::string nombre, std::string apellido, uint64_t salario, uint64_t antiguedad)
+JefeDeArea::JefeDeArea(unsigned int legajo, std::string nombre, std::string apellido, float salario, unsigned int antiguedad)
         :Empleado(legajo, nombre, apellido, salario, antiguedad)
 {
 
@@ -12,5 +12,5 @@ JefeDeArea::JefeDeArea(uint64_t legajo, std::string nombre, std::string apellido
 
 void JefeDeArea::incrementarSalario()
 {
-    this->salario = this->salario * 1.15;
+    this->salario = this->salario * 1.15f;
 }
